+++ premake5-tests.lua
@@ -134,11 +134,6 @@ local function get_simulation_app_tests()
             "standalone_examples/testing/isaacsim.benchmark.services/test_no_rendering.py",
         },
         {
-            "tests-nativepython-testing-isaacsim.simulation_app.test_external",
-            "standalone_examples/testing/isaacsim.simulation_app/test_external.py",
-            "--enable omni.kit.scripting",
-        },
-        {
             "tests-nativepython-testing-isaacsim.simulation_app.test_test_runner",
             "standalone_examples/testing/isaacsim.simulation_app/test_test_runner.py",
         },