+++ source/extensions/isaacsim.replicator.behavior/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+
+## [1.3.3] - 2026-01-22
+### Changed
+- Disable behavior test output until tests are run
+
 ## [1.3.2] - 2026-01-06
 ### Changed
 - Migrate more events to Events 2.0.